pyproject.toml
[project]
name = "leafleaf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.160.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "kivy>=2.3.1",
    "pymupdf>=1.25.2",
]

[tool.ruff]
select = ["ANN"]
